backend/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
dateparser==1.0.0
defusedxml==0.7.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
humanize==3.11.0
idna==3.2
maya==0.6.1
mysqlclient==2.0.3
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.1
pendulum==2.1.2
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
pyupbit==0.2.19
regex==2021.8.21
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
snaptime==0.2.4
soupsieve==2.2.1
sqlparse==0.4.1
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
websockets==9.1
